Require mailer config on staging and production
If we have forgotten to give the smtp config options on staging or
production, we should know at deploy time, and not have to waste time
tracking down why emails aren't sending.

This will raise an error when Rails tries to boot without the ENV
variables. On Heroku staging and production, dynos will crash
until the config variables are added (and the dynos restarted, which
happens automatically when config variables are added).