Move SMTP variable settings out of initializer
Since we require the `smtp.rb` file in the environment files, and Rails loads it
as part of its boot process, this file was loaded twice. This causes a constant
reinitialization warning.

Also, to remove the duplicaton, there is no need to check the environment,
since we require the files from the appropriate environment files.

Fixes #335