Update Rails version to 4.1.4
http://weblog.rubyonrails.org/2014/7/2/Rails_3_2_19_4_0_7_and_4_1_3_have_been_released/
http://weblog.rubyonrails.org/2014/7/2/Rails_4_0_8_and_4_1_4_have_been_released/

Fix the location of the assets version, which has changed. Use a default value
for `ENV['ASSETS_VERSION']` to make testing and other environments easier to
work with but prioritize the ENV variable so we can bust the cache in production
through changing the ENV variable without deploying new code.