Handle `dev:prime` seeds in convetional file
Rails' convention is to do DB seeding in `db/seeds.rb`. This moves the logic
from the `.rake` file to the `.rb` file.

It has the nice side-effect of being called via `rake db:seed`, in addition to
`rake dev:prime`.

Also, is allows for calls to `FactoryGirl` to be fully out-dented and not
crammed inside of nested blocks.