Automatically install Ruby version if missing
On recent projects, we've been making the Ruby setup experience better
by checking `.ruby-version`
and automatically installing the Ruby version if it's not on the machine.

This applies only to rbenv and ruby-build users,
which are the majority of thoughtbot folks.
It errors out with a friendly message for chruby or other non-rbenv users.

Our previous Bundler idempotency code also did not work
if you had a version of Bundler installed on a different Ruby version.
This commit fixes that use case.

All of this makes the situation better when a teammate upgrades
from Ruby 2.1.2 to Ruby 2.1.3 or
from Ruby 2.1.3 to Ruby 2.1.4
and you don't have that version on your machine.
All you have to do at that point is `./bin/setup` again,
just like any other "my machine is not set up properly for this project"
problem.