Don't install Foreman as a gem
Including it in the Gemfile prevents us from using the version from the
Heroku toolbelt, because the one on bin or rbenv's shims will override
/usr/bin or /usr/local/bin.

As much as possible, we think it's good to restrict bin/setup changes to
the project. The big exception for most projects is bundle install.