Turn off parameter wrapping
With this setting on, Rails will happily turn a POST to the UsersController
from this:

    { name: "Ralph" }

Into this:

    { users: { name: "Ralph" }, name: "Ralph" }

This means that `params.require(:user).permit(:name)` will not raise an error, even if the
developer forgets to nest their parameters inside a top-level `user` key.

It is enabled for JSON by default. This change turns it off for every format.