Rename development data concept to avoid db/seeds confusion
Using db/seeds for data which must be loaded in a production environment
is a standard Rails convention.  We cause confusion by using the word
"seed" elsewhere in a Rails app, where that is not the context.

The `dev:prime` task is actually providing local fake/sample data for
developers to see the app in use, but not production-required data to
actually run the application.