Update CircleCI config
The current circle.yml template does not set the env vars before running
the build, which can lead to build failures. It also doesn't set a db
setup command, which causes Circle to run `rake db:create
db:schema:load` by inference.

This PR sets the command to copy env vars from the .sample.env file post
checkout, following CircleCI's recommendation for this type of
operation: https://circleci.com/docs/manually#checkout
This allows for the env vars to be available for all steps of the build,
making it safe to run commands that load an environment, in particular
`rake`.

It also sets the database setup command to `bin/setup` to keep in line
with the Suspenders setup script.