Adds RSpec guard against truncating prod db
RSpec-rails introduced a guard against running specs in production
environment, this was introduced in version 3.3.3, in PR
rspec/rspec-rails@b913f6a.

Rails will override the environment-specific configuration from
database.yml with DATABASE_URL if it's present, so accidentally pulling
configuration from the Heroku environment is still a risk.

The guard introduced in this PR checks if DATABASE_URL is set and raises
an error.