Add mandatory environment variables
If you not set up environment variables in heroku, that are accessed using `fetch`, then the deploy will throw an error. That could by quite frustrating because you added `HOST` variable, then deploy again, and get error about `ASSET_HOST`. I think that would be great if it was present in README.