Download and use latest styleguides from thoughtbot
When users build a new project, they should have styleguides encouraging
the latest best practices available for local linting as well as through
Hound.

This change downloads the latest Ruby, JavaScript, and Sass styleguides
that we have in our thoughtbot/guides repo for immediate use. It
uncomments the relevant Hound configuration to enable their immediate
use.

Thanks to @tysongach and @teoljungberg for the idea to incorporate these
stylesheets directly, suggested in #695 and thoughtbot/guides#379.