Set development delivery_method to :file
`:test` sends the emails to the logs, and to see them one has to scroll
back through them. If we configure delivery_method to be `:file`,
mails will also be created in the `./tmp/mails/` directory

One file will be created for each email address sent to that contains
all email content delivered to that address. This allows to easily
review any and all emails sent, without searching through logs.

See:
* https://github.com/thoughtbot/suspenders/pull/521 (https://github.com/thoughtbot/suspenders/commit/8b45fd42)
* http://blog.mojotech.com/a-decade-of-rails/