Return heroku run exit code status
I noticed that migrations were failing on my deploys but heroku was
still registering it as a successful deploy. It turns out that `heroku
run` does not return the status code from the command you want to run.

Adding `--exit-code` to the `heroku run` commands will correctly return
the status code to the invoker.

Documentation: https://github.com/heroku/heroku-run