Upgrade sprockets to 4.0.0.beta3
Why:

Currently, a newly generated suspenders rails app has deprecation
warnings [per this issue]. This is due to using the `sprockets-es6`
which was not designed to be a lasting solution to transpiling ES6 to
ES5.

[per this issue]: https://github.com/thoughtbot/suspenders/issues/793

This PR:

Removes `sprockets-es6` in favor of updating `sprockets` to 4.0.0.beta3
and adding the `babel-transpiler` gem. This resolves [this issue].

[this issue]: https://github.com/thoughtbot/suspenders/issues/793