Prevent memory bloat in ActiveJob children
`ActiveJob` attaches a single `LogSubscriber` to be notified of events.
When one job enqueues another, that's treated as a child event and
stored on the same stack as the parent. If a large number of children
are enqueued that leads to R14 errors on Heroku.

The fix here is to unsubscribe all existing "enqueue" listeners and
then add a new subscriber for just that event. The original handler
is copied from the parent class using `define_method` so that it's
picked up by `attach_to`.

Related:

https://github.com/rails/rails/issues/21036
https://github.com/rails/rails/pull/5932

Before and after:

![Heroku metrics chart showing 1GB drop in memory usage after this change deployed](https://user-images.githubusercontent.com/47136/28436362-b51b3d1e-6d64-11e7-8b24-fc67927e7081.png)