Introduce a CI generator
This introduces the `suspenders:ci` generator, which sets us up with a
continuous integration configuration. In this case, Circle CI 1.0.

This is 1.0 and not 2.0 because this is a refactoring, not a feature
change.

Move the Heroku configuration to the end so that we can be sure that
`circle.yml` exists by the time it runs.