Introduce the forms generator
This introduces the `suspenders:forms` generator, which adds helpers for
working with forms. In this case, the simple_form gem.

The steps are:

1. Add the gem to the Gemfile.
2. `bundle`.
3. Create a dummy empty configuration.
4. Generate the default configuration, overwriting the dummy.

The reason for step (3) is to avoid printing a warning that we need to
generate the configuration while generating the configuration. That
warning looks like so:

```
[Simple Form] Simple Form is not configured in the application and will use the
default values. Use `rails generate simple_form:install` to generate the Simple
Form configuration.
```

This comes from `simple_form-3.5.0/lib/simple_form/railtie.rb`:

```
config.after_initialize do
  unless SimpleForm.configured?
    warn '[Simple Form] Simple Form is not configured in the application and will use the default values.' +
      ' Use `rails generate simple_form:install` to generate the Simple Form configuration.'
  end
end
```

So if we run `SimpleForm.setup` during the Rails boot, that conditional
will go false and we won't see the warning. Cleaner tests, less scary
output for the end user.