Introduce analytics generator
This introduces `suspenders:analytics`, a generator for installing app
analytics tracking. We use Segment.

This modifies an existing template to conditionalize the rendering of
the analytics partial. If the `suspenders:views` generator was not run,
do not render the partial, but do copy the partial so that the developer
can do what they please with it.