Test the suspenders:production:email generator
This is the first isolated test for a generator in this project. It
introduces a few new test helpers:

- `#with_app` takes a block and runs the block after ensuring a fresh
app.
- `#generate` runs the specified generator.
- `#match_contents` is a matcher that takes a regexp as an argument and
applies in the context for a filename expectation. It reads the file
within the dummy app and returns whether the given regexp matches the
contents of the file.
- `#rails_new` generates a Rails app.

This test runs in isolation in approximately five seconds on my
computer. This is both slow and one of the fastest tests in the repo.

We stop spring before generating in the tests using lessons learned from
3cf23ed37a94a890f7678a8ec5cd072385f93bb4.

---

While here, remove two race conditions because textbook filesystem races
irk me.