Move suspenders out of production
The suspenders gem brings in additional dependencies, not all of which
are needed in production. We only need suspenders during development --
it's a collection of Rake tasks for improving our dev speed, afterall --
so move it to the `:development` group.

Our tests need suspenders in the `:test` group. The easiest thing to do
is to hardcode that for now.

Closes #890, #891.