Added a generator for ChromeDriver
This introduces the `suspenders:chromedriver` generator option
for javascript testing if you want to use headless-chrome instead of
capybara-webkit.
It installs capybara-selenium and chromedriver-helper.
I decided to add this generator since Rails 5.2 comes with chromedriver-helper on the default setup
and it is easier to set up than capybara-webkit if you want that option.