Introduce suspenders:production:manifest
This generator copies in the `app.json` needed for Heroku, defining our
required environment variables and the app name.

Multiple generators write to `app.json`, so do not use `#template` to
copy it in. Instead, merge new JSON into the existing JSON (if any) and
write that out to the file.

It uses the RSpec type `:generator`, defined in a different branch.
Hopefully by the time this is merged, it will all just work.

---

Paired with @dimanyc.