Check for the environment, not DATABASE_URL
Suspenders was checking for the presence of `DATABASE_URL` to see if it
was running in production. This isn't so reliable in practice as
`DATABASE_URL` can be used in other environments. The spirit of this was
originally to stop you truncating a production database, which checking
for the Rails environment acheives.

https://github.com/thoughtbot/suspenders/issues/872
https://github.com/rspec/rspec-rails/pull/1383