Bump rails version to 6.0.0.rc1
In addition to bumping the Rails version, this commit makes a couple JS
changes to support Rails using webpacker by default.

The remaining changes were to get CI to pass:

- Explicitly adding the travis postgresql service might not have solved
the test failures I was having, but it did make some warnings go away.
- I don't fully understand why we need to add rack_timeout as a
development dependency and not any of the other gems, but it made Travis
happy
- I also don't fully understand why I needed to move the
`suspenders:views` generator last. The tests passed locally without
that change, and I was able to generate a working application regardless
of the order. But on CI, the generators that ran just after
`suspenders:views` were failing with errors about `flashes_helper.rb`
not existing. There is something fishy going on there, and I spent a lot
of time debugging without success. I am suspicious of bootsnap and
zeitwerk, both of which appeared in the stack trace. Since the order of
this generator doesn't matter much, I would like to go forward with this
workaround. I may come back and debug more later to see if there is a
bug somewhere.