Use Release Phase for database migration
Heroku's [Release
Phase](https://devcenter.heroku.com/articles/release-phase) feature
provides a much better migration experience for a production deploy of a
Rails application.

Without using a release phase task, here's how you migrate your database
in production:

1. Deploy your application via Git or pipeline promotion.
2. With the migration deployed, run `rake db:migrate`.
3. Immediately restart the application so that Rails is now aware of the
schema changes.

This process is awkward and can lengthen downtime in deploys. It results
in the application being restarted twice (once for the deployment of
code, and then again post-migration), and also may push application
logic that depends on the new code into production before the migration
has run or finished, which can lead to unnecessary exceptions.

There are strategies to mitigate these problems, but they generally
involve many small commits and deploys that must be carefully
orchestrated (not pushed all at once) that can be laborious and error
prone.

This commit leverages the Heroku Release Phase task to execute a
smoother migration. Here's how the process works:

1. Start your deploy with Git or a pipeline promotion
2. Heroku builds or promotes your application slug
3. Before deploying the application code to dynos, a run dyno is started
and it executes the release phase task. In this case, it's a shell
script that runs `rake db:migrate` if the application has an
`AUTO_MIGRATE_DB` environment variable set.
4. After the migration finishes, Heroku deploys the application and
restarts. If the release phase failed (migration problem), the deploy does not
proceed.

This process reduces the chance of new application code that depends on
the un-run migration being deployed and reduces the number of
application restarts from two to one. We've been using it at CommonLit
for several months now and noticed a dramatic downturn in
deployment-related downtime.

Co-authored-by: Dan Moore <dan@vaporwa.re>