Add suspenders to rails app earlier
We are seeing failures on CI for `Bundler could not find compatible
versions for gem "thor"`.

When our tests run `rails new`, that bundle installs inside the new
Rails application with thor constrained to ">= 0.20.3", "<2.0".
This started resolving to thor 1.0.0 when that version came out last
December (which is when our test suite started failing).
2019.

After running rails new, we add suspenders to the Gemfile of the new
rails application. suspenders depends on bitters, which depends on thor
"~> 0.19". This is incompatible with thor 1.0.0, which by this point is
already in the Gemfile.lock

This commit gets around the problem by adding suspenders to the Gemfile
via a template so it is there before the `bundle install`