Fix ActionMailer asset host in production.
This change was originally made in a suspenders-based app to fix a problem
we were having with images not rendering in emails. This commit replicates
the fix up-stream for future suspenders users.

The change fixes image URLs in emails by adding a protocol to the
ActionMailer asset_host setting.

While the setting is called "host", based on the examples in the Rails
documentation, it could be more accurately described as a "URL-prefix" --
all of the examples include a protocol as well as a host.

Prior to this change, image URLs were generated in emails from production
environments with a protocol-relative URL (e.g. //example.com/image.png),
and were not rendering correctly in the various mail clients we tested.