Run runner generator before profiler generator
Before this commit the profiler generator created the .sample.env. and
then the runner generator would prompt the user whether they wanted to
overwrite .sample.env. In tests this meant that the test would hang
forever waiting for user input.

After this commit the runner generator will create .sample.env, and then
the profiler generator will add to it.