Ruby 2.7.2, hardcode Rails version in tests
Bump to Ruby 2.7.2, the latest stable.

Fix test test suite: it uses the system `rails` command, which could
honestly be anything. This command supports a `_v_` flag, where `v` is a
version number; when set, it will use that version of Rails.

Modify the `suspenders` bin to also support this `_v_` option, and also
for specifying the Rails version. This implementation is lifted from the
`rails` bin.