Various CI fixes
CI doesn't have Railties

This Railties version thing is a good fix for local issues where we have
tons of gems installed, but CI has the opposite problem. The default is
probably perfect.

Actions CI, and in particular `setup-ruby`, is running into an issue with
`bundler_cache` set to `true` so remove this and add an `bundle install`
step.

Enable Postgres for CI GitHub Action

This runs [the `postgres` image from Docker Hub] as a [service] on the
test machine, which means it communicates via a network bridge. Map 5432
on the Docker instance to 5432 on the host, and export it via the
[DATABASE_URL] for Rails.

[the `postgres` image from Docker Hub]: https://hub.docker.com/_/postgres
[service]: https://docs.github.com/en/actions/guides/about-service-containers
[DATABASE_URL]: https://guides.rubyonrails.org/configuring.html#configuring-a-database

Set the `--health-...` options so GitHub Actions can perform liveness
checks.

More reading material:

- https://docs.github.com/en/actions/guides/creating-postgresql-service-containers
- https://github.com/actions/example-services/blob/master/.github/workflows/postgres-service.yml

Run tests with `DEBUG=1` so we can see more debugging output in CI
failures.

Force Postgres for `rails new` in the tests. If we don't do this, we
get:

```
LoadError: Error loading the 'postgresql' Active Record adapter. Missing a gem it depends on? pg is not part of the bundle. Add it to your Gemfile.
```

Use the doc RSpec output on CI so  we can see the names of the tests
that failed.