Skip spring in `rails new` for testing
`rails new` automatically sets up spring and "spring-ified" binstubs for
relevant commands. During testing we don't want this for two reasons:

1. We are delegating the concern of setting up spring to suspenders
`PreloaderGenerator`. To properly test this concern in isolation, we
shouldn't be creating a before state with spring already set up.

2. When `bin/rails` is "spring-ified" during testing and while calling a
generator with `bin/rails generate generator_name` the generator runs in
a new process forked by spring. For reasons not yet known, spring finds
the app root path to be the container suspenders app project directory
and when `bundle install` is subsequently run, it runs in the that
directory instead of the nested model test project directory. This
causes errors in expectations where gem installations are expected for
the model test app and not suspenders itself.

This change updates `SuspendersTestHelpers#rails_new` to run `rails new`
with the `--skip-spring` option to skip setting up spring by default for
model test app setup. It also removes `bin/spring stop` executions for
generator execution and teardown as there isn't a default expectation
that spring will be running anymore.

This bug was found in the `spec/features/advisories_spec.rb` as the
expectation that the `bundle:audit` task would be present in the
generated test app. Mike Burns <mburns@thoughtbot.com> found via
inspecting CI failures that `bundle install` was being run in the
suspenders root directory and not the model test app directory and
therefore `bundle:audit` was being set up for suspenders and not the
test app. We identified that the directory context switch happens right
when the generator is called in the test. I was able to replicate the
error locally by deleting the gem from my local gem cache and then
through further print debugging observe the context switch when spring
started before continuing the generate command.