A `gem` action
For too long we have [suffered] under the tyranny of Rails' mediocre `gem`
action! But Rails 6 took it a step too far: it [no longer] supports a
`destroy` generator.

[suffered]: https://github.com/thoughtbot/suspenders/issues/915
[no longer]:https://github.com/thoughtbot/suspenders/pull/1061#issuecomment-797764390

So it's high time we make our own. In the process, encode the rules we
like to see: keep gems alphabetical, and use the `group` block syntax
instead of the `:group` argument.

When inserting a gem, we have to consider two cases: do we have a group
specification? If we don't have a group name, insert it alphabetically
before the first non-indented `gem "..."` declaration. If we do have a
group name, find the block for that group and insert it alphabetically.
But there's also the chance thar we are the first to add this group, in
which case we should add the group with the `gem` line at the end.

When removing a group, we again consider the groups case. If there is no
group specification, remove the first matching non-indented `gem "..."`
declaration. If we have a group spec, find the group block in the
Gemfile and remove it. If there is no such group block, do nothing.

Closes #915.