Make Suspenders abort when something goes wrong
** What changed?

`bin/suspenders` is now halting on error. Not halting on error was a
huge source of confusion for Suspenders users. Some examples of issues
we are resolving here:

- If using the wrong Ruby version, Suspenders will fail loudly.
Currently, Suspenders is limited to a single Ruby version (we'll make
the Ruby version flexible in another PR.) Example:

```
      create  storage/.keep
      create  tmp/storage
      create  tmp/storage/.keep
      remove  config/initializers/cors.rb
      remove  config/initializers/new_framework_defaults_6_1.rb
       force  Gemfile
         run  bundle install
Your Ruby version is 2.7.5, but your Gemfile specified 2.7.4
```

No more output. The user will instantly know something is wrong and
will attempt to fix the error.

- If Postgres is not running, Suspenders will fail when trying to
create the database

- If a generator fails because of some unmet requirement, Suspenders
will halt as well (recall that every individual generator runs within
a subprocess.)

** Changelog

- Override thor's `exit_on_failure?` to `true` on `AppGenerator`.
`AppGenerator` is the class that extends
`Rails::Generators::AppGenerator` and runs our own version of `rails
new` through `bin/suspenders`.

- Override thor's `exit_on_failure?` to `true` on
`Suspenders::Generators::Base` (the base class of our individual
generators.)

- Make `Rails::Generators::AppGenerator#bundle_command` fail on error.
Unfortunately, it is not hooked up to thor's default error handling
mechanism, so it won't leverage Thor's `exit_on_failure` setting.

- Override the default value of `skip_spring` to `true`. Rails will
still try to generate spring binstubs when finishing up the
application, which would make `bin/suspenders` fail given our newly
introduced `exit_on_failure` setting.

- Avoid the `spring stop` command from halting `bin/suspenders` when
`spring` is not installed or when it calls a binstub that can't find
the real `spring`.

- Add unit tests for the new settings. `exit_on_failure` is a
`:nodoc:` in Thor, and it's unfortunate they don't provide an official
setting for that. If it ever breaks, we'll need to know.

- Don't run the stylesheets generator when `--api` is set. That avoids
the stylesheet generator (and thus `bin/suspenders`) from failing with
our new settings.

- Rename `DEBUG` => `SHOW_DEBUG` because `DEBUG` is already taken by
Ruby and prints out too many symbols.