Support Rails 7 and Ruby 3 (#1106)
This is a major upgrade that aims to make it so generated Rails
applications with suspenders are running on Rails version 7 and Ruby
version 3. We took the approach of making the least amount of changes
possible.

We were unable to support the latest version of Ruby because of a
[failing build][]. A future commit could explore how to solve for this,
but for now, we decided to just stick with version `3.0.5`. We settled
on `3.0.5` and not `3.0.0` in an effort to suppress parser warnings.

The biggest change is that we drop support for deprecated CSS tooling by
removing [Bourbon][] and introduce [PostCSS][] and
[cssbundling-rails][]. We also drop [Bitters][] as a runtime dependency.

Add PostCSS Normalize
---------------------

In order to use [normalize][], we need to use [PostCSS Normalize][]
since we are using [cssbundling-rails][].

We need to move the invocation of the `suspenders:stylesheet_base`
to the `leftovers` method because this method is invoked _after_
`finish_template` is called. We do this because we were running into
issues where the call to `rails css:install:postcss` had not yet run,
but our generator assumes that command has already run. The method must
be named `leftovers` since that's what the [finish_template][] method
expects.

Additionally, we only call this generator if the caller has not
passed anything to the `--css` option when generating a new app. This is
because Rails handles other css frameworks like `bootstrap` and
`tailwind` differently.

Finally, we move the invocation of `outro` to the `leftovers` method
since it ensures the message will be the last thing to be printed to the
console.

Add bin/yarn
------------

Now that we have moved off of [Webpacker][], the [bin/yarn][] binstub is
no longer generated. Because of this, we need to manually have it added
during a new app build.

Introduce CalVer
----------------

Because this upgrade was a breaking change, we thought it was a good
time to switch to [CalVer][]. Since this project is composed of other
projects, every release is a potential breaking change. Using the date
as the `major` version encodes this risk.

Improve console output
----------------------

Ensures the invocation of all generator methods happens _before_ we
print the outro. This is because a generator invokes all methods in the
class in the order they are declared. Specifically, this meant invoking
`generate_views` in the `suspenders_customization` method. Prior to
this, `generate_views` was invoked after the `outro` was printed.

[Webpacker]: https://github.com/rails/webpacker
[failing build]: https://github.com/thoughtbot/suspenders/actions/runs/3418310707
[Bourbon]: https://github.com/thoughtbot/bourbon
[PostCSS]: https://postcss.org
[cssbundling-rails]: https://github.com/rails/cssbundling-rails
[Bitters]: https://github.com/thoughtbot/bitters
[normalize]: https://necolas.github.io/normalize.css/
[PostCSS Normalize]: https://github.com/csstools/postcss-normalize
[finish_template]: https://github.com/rails/rails/blob/d5124b2522130785378238ba714029ce8ef9de97/railties/lib/rails/generators/rails/plugin/plugin_generator.rb#L283-L285
[bin/yarn]: https://github.com/rails/webpacker/blob/master/lib/install/bin/yarn
[CalVer]: https://calver.org

Co-authored-by: Aji Slater <aji@thoughtbot.com>
Co-authored-by: Eric Milford <ericmilford@gmail.com>
Co-authored-by: Stefanni Brasil <stefanni@thoughtbot.com>