Lock standard version to fix failures
The version of `node` isn't the issue, but I __think__ it's standard.

I was wondering why our Gemfile.lock file was ignored, so I looked at
the [commit][1] that introduced that change. It mentioned that:

> suspenders as a gem relies on version ranges for its dependencies
defined in its gemspec

This made me realize that the one failure we get over and over involves
standard.

It makes sense that CI was passing in the past, because standard was at
a different version.

This commit first locked the version to `1.9.0`, which is the version we
were using during the last successful build. From there, I stepped
through each [release][2] until I got a failure, which was at `1.26.0`.
This lead to me pessimistically lock the value so it will resolved to
`1.25.5`, which is the latest version to **not** break CI.

The fact that my tests were passing locally was probably because my
`Gemfile.lock` reflected an older standard version, but since deleting
`Gemfile.lock` I'm getting failures.

[1]: https://github.com/thoughtbot/suspenders/commit/539b5c6e7fe68ee6ccad168ea6e267c7f12ae9bd
[2]: https://github.com/standardrb/standard/releases