Revert changes to node version declaration
This was a false flag. I thought my tests were passing locally because I
was on a different version of node, but it turns out my `Gemfile.lock`
was set to an older version of standard that did not break tests.

Once I removed the `Gemfile.lock` file and ran `./bin/setup`, I was able
to replicate the failure.