Allow suspenders to be used with a range of Ruby versions (#1129)
Suspenders locks the user in a particular Ruby version (3.0.5 at the
time of writing), a restriction that can sometimes be annoying. This
PR enables Suspenders to be used with a range of safe Ruby versions,
initially from 3.0.5 to 3.2.2

Contrary to the Rails version, the Ruby version can be more flexible
if our gems are compatible. Major Rails versions, on the other hand,
tend to be a bit more disruptive and may require deeper changes to
Suspenders. [The last main breaking
change](https://www.ruby-lang.org/en/news/2019/12/12/separation-of-positional-and-keyword-arguments-in-ruby-3-0/)
was on Ruby 3.0, but we're already past that point.

`.ruby-versions` was the source of truth for the canonical Ruby
version, but in this PR it represents the maximum version.
Unfortunately, it is still needed by CI so we can't delete it. Note
that even without such a file, rubybems validates the runtime Ruby
version through the gemspec.

Closes https://github.com/thoughtbot/suspenders/issues/1073