Suspenders 3.0.0
In an effort to improve the developer experience and reduce technical
debt, we replace the existing project with the output from running:

```
rails plugin new suspenders
```

This will allow us to use existing [testing helpers][] provided by Rails.
Because of this, we deliberately switch from RSpec to MiniTest.

Additionally, we move away from [calver][]. Prior to that the latest major
version was `1`. This feels like a good opportunity to go back to `semver` and
start fresh with `3.0.0`

[testing helpers]: https://edgeguides.rubyonrails.org/generators.html#testing-generators
[calver]: https://calver.org