Introduce test helpers
All generators will affect the "test" app located in `test/dummy`.
Although the "test" app contains most files for a Rails application, it
does not contain everything. This means during the setup phase of our
tests, we'll need to create files to ensure the generator is working
correctly.

These helpers aim to improve the developer experience by creating a
simple API to create and delete files in the "test" app.

It should be noted that any files and directories created in the setup
won't be removed automatically after each test, so we introduce
additional helpers to be called during the [teardown][] callback.

[teardown]: https://github.com/rails/rails/blob/00dfa109d6a5f4fb13d745f83d5cc779eba3b352/railties/lib/rails/generators/testing/setup_and_teardown.rb#L13