Introduce `suspenders:rake` generator
Add necessary files to make the [plugin][] an [engine][], which
automatically loads Rake tasks located in `lib/tasks`. This means when
the `suspenders` gem is installed, the consumer can run `bundle exec
rake suspenders:rake`, and any future tasks.

Because `suspenders:lint` and `suspenders:advisories` may not
necessarily have been invoked, we need to check if those gems are
installed.

[plugin]: https://guides.rubyonrails.org/plugins.html
[engine]: https://guides.rubyonrails.org/engines.html

Co-authored-by: Mike Burns <mburns@thoughtbot.com>